requirements.txt
# Video Stabilization System - Requirements

# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.0
opencv-contrib-python>=4.5.0

# Configuration management
pyyaml>=6.0

# Video processing utilities
imageio>=2.9.0
imageio-ffmpeg>=0.4.0

# Progress bars
tqdm>=4.62.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Linting and formatting (development)
black>=22.0.0
flake8>=4.0.0
pylint>=2.12.0

# Documentation
sphinx>=4.3.0
sphinx-rtd-theme>=1.0.0

# Jupyter notebook support (optional)
jupyter>=1.0.0
matplotlib>=3.5.0

# Streamlit web app
streamlit>=1.28.0
